Fixes collapsible elements.
The classes used for the collapsible elements didn't work as expected.
The new ones do, at least in my tests. There's still a problem in the
'Assignment text' collapsible element in ecr_view.pt and ecr_grade.pt
but since that comes from the eca_view_head macro defined in
ECAssignmentBox the problem is a little bit harder to fix.