Fix buggy handling of "no submissions" case.
The code for the case that the referenced assignment box doesn't contain
submissions eligible for review was just wrong.
The new code still doesn't properly abort the creation of the review box
with an error, but the fact that there's something wrong and gives a few
pointers on how to fix this.