Change the return value of 'availableReviews()'.
In order to be able to display the review for each users own submission
in a special way for him, knowledge over which review belongs to which
user's submission is required. Therefore ECReviewBox#availableReviews()
now returns a dictionary where the keys are usernames.
The 'view_available_reviews.cpt' template had to be changed to work way
of returning the available reviews.
In addition to that no randomization is done anymore, because the a
dictionary has an undefined order anyway.