"Fix" the parent link on the 'Reviews' tab.
Since an ECRB will usually reside in an ECFolder there is normally no
action corresponding to the 'Reviews' action on ECRB's containter.
Therefore the 'up one level' link just points to the default parent and
the default template is used to render it.
This may not be desired but is better than linking to a nonexistent URL
as it is the case before this patch.